sign up form refactored to atomic desgin
