+++ components/organisms/LoginForm.js
@@ -1,7 +1,8 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { StyleSheet, Text } from 'react-native';
 import { useNavigation } from '@react-navigation/core';
 import { login } from '../../services/user';
+
 import TextField from '../atoms/TextField';
 import SecureField from '../atoms/SecureField';
 import Form from '../molecules/Form';
@@ -33,13 +34,11 @@ const LoginForm = () => {
             placeholder='Email'
             value={ email }
             onChangeText={text => setEmail(text)}
-            style={styles.input}
           />
           <SecureField
             placeholder='Password'
             value={ password }
             onChangeText={password => setPassword(password) }
-            style={styles.input}
           />
 
           <Button onPress={handleLogin} text={'Login'} />
